Remove parser package, and create a single parser module
